backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
reportlab==4.0.7
pandas==2.1.3
requests==2.31.0
tenacity==8.2.3
stripe==7.8.0
python-dotenv==1.0.0
email-validator==2.3.0